Making adding of the stream component more conservative
Note, this changes the workflow ever so slightly. The issue, is that in
some cases (when brush influence was on, etc), the VertexInstanceStream
component would get added to objects which were not being painted on.
Now these are only added when absolutely necessary, but the downside is
that you can no longer use ‘Show Vertex Data’ without the tool being
active and the stream being already added.

This means that you have to paint before you’ll see the vertex data,
which is not very nice. So, to at least make this better, there’s now
an “Add Stream” button next to the preview toggle. Clicking it will add
the stream and start showing the vertex data if preview mode is on.
While I was there, I also added a “Remove Unused” if none of the
streams have data. This makes it a bit easier to cleanup as well.