Refactoring, bug fixing
- Fixed a bug which would cause the preview shader to get repeatedly
put back on your model instead of your actual shader
- Refactored Bake panel, which is now called utilities. You can now add
utilities to the panel in the same manner as custom brushes; external
classes that plugin to the frame work and get access to the job data
- Converted all old ‘bake’ functions into the new plugin interface
- GUI is now based on rollouts; any header in the UI can be clicked on
to show/hide the area below it. This looks a lot nicer than foldouts,
but serves the same general use case.