multiply of mouse position only on OSX
I tested this on 5.4 as well as my specific patch release on OS X and I
seem to need the 2x for things to line up. But on Windows, it doesn’t
need the 2x. If anyone is seeing the mouse position not line up in OS
X, please, please let me know.