Fixing mouse position on rentina displays on 5.4
So Unity decided to add a value thats “multiply your mouse position by
this, or it’ll be off on some monitors” to 5.4. If your going to
require us to multiply mouse position every time we use it, why not
just send us it in the right space in the first place?